fix(windows): suppress MSVC 18 experimental coroutine deprecation + signed/unsigned cast
- Add _SILENCE_EXPERIMENTAL_COROUTINE_DEPRECATION_WARNINGS to CMakeLists.txt
  for flutter_local_notifications_windows plugin (STL1011 with VS 2025)
- Cast CW_USEDEFAULT to unsigned int to fix C4245 warning-as-error

+++ windows/runner/main.cpp
@@ -54,7 +54,8 @@ int APIENTRY wWinMain(_In_ HINSTANCE instance, _In_opt_ HINSTANCE prev,
   FlutterWindow window(project);
   // Let Windows choose the initial position (centered/cascaded).
   // Size matches Flutter's WindowOptions(size: Size(1280, 820)).
-  Win32Window::Point origin(CW_USEDEFAULT, CW_USEDEFAULT);
+  Win32Window::Point origin(static_cast<unsigned int>(CW_USEDEFAULT),
+                            static_cast<unsigned int>(CW_USEDEFAULT));
   Win32Window::Size size(1280, 820);
   if (!window.Create(L"Candela", origin, size)) {
     return EXIT_FAILURE;
